+++ src/main/java/com/ucaldas/posgrados/Controller/EgresosOtrosServDocentesController.java
@@ -73,7 +73,7 @@ public class EgresosOtrosServDocentesController {
             // Guardar el Presupuesto actualizado
             presupuestoRepository.save(presupuesto.get());
 
-            return "Egreso de otros servicios docentes guardado";
+            return "OK";
         } else {
             return "Error: Presupuesto no encontrado";
         }
@@ -115,7 +115,7 @@ public class EgresosOtrosServDocentesController {
             egresoOtrosServDocenteRepository.save(egresosOtrosServDocentesActualizado);
 
             presupuestoController.actualizarEgresosProgramaTotales(idPresupuesto, valorNuevo, valorAnterior);
-            return "Egreso de otros servicios docentes actualizado";
+            return "OK";
         } else {
             return "Error: Egreso de otros servicios docentes no encontrado";
         }
@@ -125,12 +125,16 @@ public class EgresosOtrosServDocentesController {
     public @ResponseBody String eliminar(@RequestParam int id) {
 
         Optional<EgresosOtrosServDocentes> egreso = egresoOtrosServDocenteRepository.findById(id);
+
+        if (!egreso.isPresent()) {
+            return "Error: Egreso de otros servicios docentes no encontrado";
+        }
         int idPresupuesto = egreso.get().getPresupuesto().getId();
         double valorAnterior = egreso.get().getValorTotal();
 
         presupuestoController.actualizarEgresosProgramaTotales(idPresupuesto, 0, valorAnterior);
         egresoOtrosServDocenteRepository.deleteById(id);
-        return "Egreso de otros servicios docentes eliminado";
+        return "OK";
     }
 
 }
